pyproject.toml
[tool]
[tool.poetry]
name = "aimet-ml"
version = "0.1.0"
homepage = "https://github.com/pasitj-aimet/aimet-ml"
description = "Python package of frequently used modules for ML developments in AIMET.."
authors = ["Pasit Jakkrawankul <pasit.j@aimet.tech>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "aimet_ml" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"

black  = { version = "^23.7", optional = true}
isort  = { version = "^5.12.0", optional = true}
flake8  = { version = "^6.1.0", optional = true}
flake8-docstrings = { version = "^1.7.0", optional = true }
mypy = {version = "^1.5.1", optional = true}
pytest  = { version = "^7.4.0", optional = true}
pytest-cov  = { version = "^4.1.0", optional = true}
tox  = { version = "^4.10.0", optional = true}
virtualenv  = { version = "^20.24.3", optional = true}
pip  = { version = "^23.2.1", optional = true}
mkdocs  = { version = "^1.5.2", optional = true}
mkdocs-include-markdown-plugin  = { version = "^6.0.0", optional = true}
mkdocs-material  = { version = "^9.2.3", optional = true}
mkdocstrings  = { version = "^0.22.0", optional = true}
mkdocs-material-extensions  = { version = "^1.1.1", optional = true}
mkdocstrings-python = { version = "^1.5.2", optional = true}
twine  = { version = "^4.0.2", optional = true}
mkdocs-autorefs = {version = "^0.5.0", optional = true}
pre-commit = {version = "^3.3.3", optional = true}
toml = {version = "^0.10.2", optional = true}
bump2version = {version = "^1.0.1", optional = true}
ffmpeg_python = "^0.2.0"
matplotlib = "^3.7.2"
numpy = "^1.24.4"
opencv_python = "^4.8.0.76"
pandas = "^2.0.3"
pydub = "^0.25.1"
librosa = "^0.10.1"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
transformers = "^4.32.0"



[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
